Generalize vehicle explosion, make all vehicles use shallow water, add explosion depth tolerance

+++ TombEngine/Objects/TR4/Vehicles/motorbike_info.h
@@ -8,7 +8,7 @@ struct MotorbikeInfo {
 	short momentumAngle;
 	short extraRotation;
 	short wallShiftRotation;
-	int bikeTurn;
+	short bikeTurn;
 	int pitch;
 	short flags;
 	short lightPower;
